+++ sql/message/search.sql
@@ -1,6 +1,10 @@
 SELECT
-    *
+    date,
+    host,
+    program,
+    message
 FROM
     message
 WHERE
-    message LIKE ?;
+    message LIKE ?
+ORDER BY rowid;
