+++ src/lib.rs
@@ -1,6 +1,5 @@
-pub mod api_server;
+pub mod api;
 pub mod cli;
-pub mod fetch_server;
-pub mod message;
-pub mod message_api;
-pub mod processing_server;
+pub mod database;
+pub mod model;
+pub mod server;
