+++ src/model/mod.rs
@@ -1,0 +1,2 @@
+pub mod message;
+pub mod user;
