+++ src/server/mod.rs
@@ -1,0 +1,3 @@
+pub mod api;
+pub mod fetch;
+pub mod processing;
